feat: Implement heritage data fetching and error handling in PlaceDetailPage
- Added new `heritage` entity files for API and model definitions, including the `fetchHeritageById` function to retrieve heritage details based on ID and location.
- Integrated `heritageItemQueryOptions` for managing queries with React Query, enhancing data fetching capabilities.
- Refactored `PlaceDetailPage` to utilize the new query structure, replacing previous data fetching logic with a more robust error handling mechanism using `ErrorBoundary` and `Suspense`.
- Introduced `ErrorFallback` component for user-friendly error handling during data retrieval.
- These changes improve the overall architecture and user experience by ensuring accurate data fetching and error management.
