+++ src/pages/place-detail/api/heritageItem.query.ts
@@ -1,0 +1,8 @@
+import { fetchHeritageById } from '@/entities/heritage';
+import { queryOptions } from '@tanstack/react-query';
+
+export const heritageItemQueryOptions = (id: string, lat: number, lng: number) =>
+  queryOptions({
+    queryKey: ['heritage', id, lat, lng],
+    queryFn: () => fetchHeritageById(id, lat, lng),
+  });
