+++ src/shared/ui/index.ts
@@ -4,3 +4,4 @@ export { QRScanner } from './QRScanner';
 export { Nav } from './Nav';
 export { FixedBottom } from './FixedBottom';
 export { NoResult } from './NoResult';
+export { ErrorFallback } from './ErrorFallback';
